Fix history view context menu
The entries "Compare with working tree", 
"Compare with each other" and "Open"were shown not only 
for a commit but also for a file in the list of files contained 
in a commit. However, they opened a compare editor 
for the resource which is the input of the 
history view.

Change-Id: I6305e47b7a154920f2ef38292ff649e23bcb6f42
Signed-off-by: Stefan Lay <stefan.lay@sap.com>