Implement CommitFileRevision.equals() and hashCode() so editors can be reused
When editors are asked to be opened in the Eclipse workbench, the
default behaviour is to simply activate an existing editor if an
existing editor shares the input with the input that is to be
opened. The contract for comparing editor inputs is via the
equals(Object) method. For a FileRevisionEditorInput, this
is implemented by checking the equality between the two inputs'
underlying file revisions. As such, CommitFileRevision should
implement equals(Object) (and hashCode() for consistency) so that
the same editor can be reused instead of having multiple editors
open on a given revision.

Change-Id: I2abeef934b21986979f84f9836e0bf2b44eff41f
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>