'Repositories' view should not show scroll bars unless it has to
The view's code was arbitrarily introducing a TreeColumn to the
view's tree which had a width of 700. This meant that when the
view's width became less than 700, it would have to show scroll
bars to accomodate for this. However, it should actually only be
showing scroll bars if it has to and not because of some arbitrary
value has been surpassed.

The worst part is that using columns means that the text can only
go as far as what is dictated by its width. Since the view does not
have any column headers displayed, the user actually has no way of
resizing the column to view all of the content. The fix is to just
not use any columns and to let the tree itself determine whether it
needs to show scroll bars or not based on whether it is currently
displaying content that exceeds its own width.

Change-Id: Ib1da786859e0368816e229f03d9d6d1d30ec962b