Fix broken 'Show In' feature for 'Git Repositories' view
Code was added to make the 'Git Repositories' view show up in the
'Show In' menu but no code was introduced in the view to actually
react to the request. As is, the view will simply be shown without
actually having anything selected. The fix is to implement the
IShowInTarget interface so that the request can be handled.

Change-Id: I64e11f856c25970c07e03dc3365e24f7e1733ede