Prevent leaking of SWT menus in the 'Git Repositories' view
New instances of SWT menus are currently being instantiated and set
on the tree control every time a user brings up the context menu in
the 'Git Repositories' view. When the tree is disposed, the tree's
menu is also disposed. However, as new menus are instantiated and
set every time, those menus that have been replaced are now lying
around and will only be disposed when the shell itself is disposed.

Change-Id: Ia77bf8cbd8cd3ef27e536821dd979e0882926f44