Improve Repositories View performance
RepositoriesView now uses the shared Repository instances provided by
RepositoryCache instead of creating own instances.
This enables RepositoriesView to listen to repository change events.
The auto refresh job of RepositoriesView is no longer needed.
Activator.RCS was enhanced to scan all repositories contained in
RepositoryCache. This was required because RepositoriesView is the first
client of RepositoryCache that uses Repository instances which might
not have a project in the workspace.

Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>