NullPointerException in Git History View
This occurs because these menu actions are implemented
using a preference change listener. When the page is disposed,
the corresponding actions are not disposed by the history framework,
so the registered actions keep listening for the preference changes.
To reproduce, open a Git History, then open "History" on some non-
Git controlled object in order to dispose the page. Open again a Git
History page, then click on the view menu -> Wrap comments (or any
other action there).
The fix actively disposes the registered actions when the page is disposed.

Bug: 316277

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>