Better handling of invalid WindowCache Configuration
Currently, if an invalid value for the WindowCache
configuration is stored (e.g. "2 k" for window size), then start-up of
EGit fails after the next IDE restart.
Furthermore, no suitable error message is shown when this value
is stored in the preference page.
The fix makes the startup failsafe by logging an error if the config
is wrong but starting the cache with the default config. Also, the
preference page shows an error message if a wrong value is
entered.

Bug: 316761
Change-Id: If06dec45ff521b4bfaa21d91fc83f541eb3bde28
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>