Git Repositories View: UI tests
This increases the overall test coverage for the EGit UI plug-in
from around 20 to around 41 percent. Also explicitly specify
the test suite for maven build so that it picks up the right
test suite.

While writing these tests I found a number of bugs. 11 of 28 tests
had to be set to ignore so that the tests are green. Fixes will be
submitted separately and tests will be re-activated accordingly.

Change-Id: I8c0f5016f07c110a718536b10dd577219711c7e9
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>