Add repository state to Git Project Label Decorations
With the advent of merge capabilities, we should
make the repository state (mostly merge, rebase)
 visible in the Git label decorationsfor projects.
Instead of adding another tag, we re-use the
"repository name" tag and append the repository state separated by 
a pipe similar to the behavior in msysgit. 
This allows to elegantly treat the "SAFE" state that we don't
want to show explicitly: we just add nothing to the repository name.
This avoids ugly formatting issues which we would run into
if using a separate tag: unless we would want to always show
the state (even if it is SAFE) we would introduce ugly spaces
into the label decoration in case the state is SAFE (and thus the
displayed value is null or an empty String). The alternative of howing 
the state always does not appear to be the right thing, either, as it
would pollute the label decoration with un-needed information in the
99 percent case.

Change-Id: Ia20a9b60238077acf9e85c22c7a7958cbb50bfd9
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>