Minor fix for UI tests to get the correct button on GTK
On GTK, the default behavior is to show the cancel button before the OK button in
all dialogs. In the tests we should not assume the order of the buttons but try
to match by the button text. In order to run the tests on GTK, it is neccessary to identify
the right button as all tests currently "click" the Cancel button when trying to
open the Repository View.

Change-Id: Icc767830e398a3144bb17a61a286eebcb5113aa8
Signed-off-by: Benjamin Muskalla <bmuskalla@eclipsesource.com>
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>