Fix GitCloneWizardTest.updatesParameterFieldsInImportDialogWhenURIIsUpdated
The test expected an specific error message due to an imcomplete URI.
However, the checking code did not consider the details in why the URI
was incomplete.

Now, if you leave out the host, the GUI will (again) tell you that.

Also add a launcher with more memory assigned to avoid failures due to
out-of-memory.

Change-Id: Ib22cc352e221f74d344c2241abaee6e6628fbb41
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>