Switch to 3.5 trace infrastructure
This introduces the OSGI 3.5 trace infrastructure and replaces the
current minimal implementation with the proper one which
provides powerful features (automatic addition of method names,
thread name, line numbers and so on).
The next steps will be to instrument our coding and to provide
a UI for dynamically setting the debug options during runtime
and to add some tests.

Change-Id: I3ca833407762f122db329eec4da0083aa64fa918
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>