Trigger a merge commit
Allows to commit when merge conflicts are resolved.
To avoid unnecessary complexity file selection is switched
off in the commit dialog. Workspace selection is not
taken into account.

Uses JGit's commit command which handles determination
of parents and resetting repo state.

The merge message is preset in the dialog. No addititional
hints are presented (like when using the command line).

Bug: 318439
Change-Id: I855903ba2f620aaadf38d4de7268cffccf38b0a9
Signed-off-by: Stefan Lay <stefan.lay@sap.com>