History View: build menu programmatically
It turned out that using the "global" (workbench) selection
for building the context menu in the history view is not
feasible, as the GenericHistoryView does not handle
registration of the selection providers properly (see
bug 322751).
Thus we build the context menu programmatically again
(using commands instead of actions) and remove the 
menu contributions from the plugin.xml.
The changes in the test coding were needed to fix
issues with failing tests (some of them, but not all
being related to the changes in this commit).

Some UI tests are still failing due to bug 322868.

Bug: 321516
Bug: 322751
Change-Id: I1375afc763f15b0e5a1b932bbb1ae24694ea6428
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>