Pass WorkingTreeOptions to JGit's iterators
JGit recently learned to pass a WorkingTreeOptions to its iterators
to control how CRLF conversion should be performed when computing
the ObjectId for a file entry.  Pass the default instance for now
just to fix the compile errors, but we should come back and change
these to honor the underlying Git repository's configuration.

Change-Id: I4eb363c163e89c5bc6faa5d8c1ad9222450d4532
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>