Use standard Changeset Icon for Synchronize View
We should try to avoid confusion with different concepts. Changesets
should have the same icon as they do in other team providers and not
have the repository look&feel. Seems someone already had the idea
because the icon is already there, just the right usage in the label
provider was missing.

Bug: 323475
Change-Id: I38a541088eb2cf145686a47ee9f8dc50afe95f7e
Signed-off-by: Benjamin Muskalla <bmuskalla@eclipsesource.com>