Clear RepositoryCache after each test to fix core tests
With this fix the core test suite is running also on Mac. Reason for the
failures before was the usage of out-dated Repository instances
contained in RepositoryCache which caused spurious failures when
running the full test suite while running tests individually always
succeeded.

A follow-up fix should care for the life cycle of Repository objects
in RepositoryCache.

Change-Id: I1f16a9de48435bd98e2520e34e49bc1ca6893f14
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>