Add commit hash to history view table
When comparing two revisions from the history the compare view
shows the commit hash next the file name but from the history view
table it was not possible to identify revisions by commit-id.
To find out which revision is which you needed to select a single
revision from the history view and then check the detail pane for
that commit.

Bug: 327291
Change-Id: I073ed81ca022fc22779accff75fc56d7ffb98dfc
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>