[historyView] Fix tag handling when "moving" tags
This was found while analyzing the bug below. While the original issue
of that bug is already fixed with

http://egit.eclipse.org/r/#change,1807

we still have an issue with the code used for "moving" the tags in the
history view. The symptom is that a request to move an existing tag
into the currently selected commit is simply ignored.

Bug: 327971
Change-Id: I3353ab88b1ca17bb943576ec8144141e6a8aa5fd
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>