Add cheat sheet for clone wizard
This change contains a cheat sheet for the clone wizard to support
users that are new to EGit. It can be found in Eclipse as follows:
Help > Cheat Sheets... > Team/Git > Cloning a Git Repository.
Furthermore a help context for the clone wizard has been added that
can be launched by clicking on the help icon (?) from within the
wizard. The help context also provides access to the cheat sheet.

Change-Id: If6fd730154a2e59daeff89d96cab2006718d8d9f
Signed-off-by: Philipp Thun <philipp.thun@sap.com>