Improve presentation in project sharing dialog
Improve presentation of relative directories when there are multiple
choices for the .git directory. In particular the project was
presented as '/' instead of '.' for the case in the subject line of
this issue. It does not fix Bug 297163, but perhaps makes it less
likely that the user chooses the wrong .git directory.

We also add tests for this case.

Bug: 297163
Change-Id: Ic2475b46127ddcc977392beb48749a123bd7b800
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>