Revert I33e64c99
This was a mistake that was missed due to history reasons.

"The first /r/ tells our Apache to redirect the request to Gerrit.
The second /r/ tells Gerrit that the thing following is a Git SHA-1
and it should try to locate the changes that use that commit object.
Nothing I can easily do about it now. The second /r/ is historical
and comes from Gerrit 1.x days."

Change-Id: I315cd56bcd474d95ad82d285bcc92fd6e65d4816
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>