Add "quick" compare and use it in history view by default
When clicking "Compare with each other" after selecting exactly two
commits while showing a project, folder, or repository in the history
view, we open a "quick" compare (using the Eclipse compare
infrastructure) to show the structural differences instead of the
Git Tree Compare view. A new menu action is added to allow to open the
Tree Compare.

Change-Id: I411e290131647c969b96df71eb9c7c0f899b7c6b
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>