Fix HistoryView tests after commit 611459632b77
Due to the asynchronous updates, we have to add a few waits to make
the tests work again. In the longer run, we should probably try and
decouple the commit table, the message viewer, and the file list
updates completely from each other by spawning jobs. Then the tests
should join with these jobs to obtain proper synchronization.

Change-Id: I8b872d82cc1da64a4db884b3e3848ca57e824645
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>