Proper result dialog when pulling from multiple Repositories
The dialog presents a list of Repositories and an overview about fetch
and update (merge/rebase) results for each Repository. Users can drill
down and see detailed information (i.e. the "normal" PullResultDialog)
by selecting a line in the list and hitting a "Details" button.

This also enables pull for multiple Repositories in the Git
Repositories View.

Bug: 340780
Change-Id: I8d4aaf3f3f0b5f47cf0db50b4588a2676bb8ccc2
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>