Use workspace file encoding in commit dialog diff
When viewing the diff (from the commit dialog) of a file that has a
different encoding than the workspace encoding, it contains incorrect
changes.

One way to reproduce this problem:

- Set the Eclipse workspace encoding to ISO-8859-1.
- Create an XML file that has an explicit UTF-8 encoding set (either
  through an encoding attribute in the XML or as an explicit property on
  the resource). Add some non-ASCII characters in the file.
- Commit it.
- Change some unrelated lines in the file.
- Open the commit dialog.
- View the diff of the file.

Before this change, the diff would show changes in the lines with
non-ASCII characters, even though they didn't change. The reason is that
the encoding of nextFile is correct because it is from the workspace.
But it's wrong for baseFile because it's read as a blob from the repo,
and the diff view therefore falls back to the wrong workspace encoding.

After this change, the encoding from the IFile is also used to decode
the previous file contents, and the diff is correct.

Change-Id: I58ba71d30d7dd47068d2b212737896f6c1b9843a