Allow dragging of commits from History view
User are now able to drag & drop a commit into other applications (eg.
text editor or file explorer) and will get the patch for this specific
commit. Furthermore this allows better integration with existing plugins
like Mylyn - drag and drop a commit onto the task editor to attach the
commit as patch to the task.

Change-Id: Ia27cb669230ec867deabc354cc657ab4721e7900
Signed-off-by: Benjamin Muskalla <benjamin.muskalla@tasktop.com>