Don't ask Compare to open two editors for selected IFiles
The 'Git Tree Compare' view had incorrect logic that caused two
editor open requests to be sent to Compare. This meant that when
the 'Reuse Compare Editor' option was selected, the editor would
get closed and when it wasn't selected, two compare editors
would get opened. Removing the first request and letting the
code cascade down to the request at the end of the method fixes
this problem.

Bug: 345631
Change-Id: I6b300a18e56af6059a34d751bb5c63abcb362c61
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>