Ignore null returns for getAdapter(Class)
When an element is retrieved from the selection service,
getAdapter(IProject.class) is invoked on it to determine if the
selection can be adapted to become a project. If that cannot be
satisfied, the code should be allowed to proceed normally and
treat the object as it is in the event that the currently
selected object is of a type that supports the switching of
branches.

Bug: 361556
Change-Id: Iee98076ae1bfc050263537decc1ce442448cc26d
Signed-off-by: Remy Suen <remysuen@ca.ibm.com>