Do not set a filter if an empty string is present
As the compare editor input uses the input resources for
determining what it should filter to show, its filter may
potentially be an empty string if the project is both the parent
of the .git folder as well as the selected root resource for the
compare editor to display. If this scenario is satisfied, simply
skip the discovered empty string and don't bother setting a
filter on the walk.

Bug: 359308
Change-Id: Ib303ff689b6a7a70f5aa4ba6fa986a94f7997f87
Signed-off-by: Remy Suen <remysuen@ca.ibm.com>