[historyView] Add preferences for max branch and tag length
This controls how branch and tag labels in the table are
truncated with an ellipsis when the specified length is exceeded.

This breaks the "Views" preference page into two separate pages,
one for the History view and one for the Synchronize view.

The initial default for both preferences is 15 which is
larger than the previous internal value of 12.

Bug: 341593
Bug: 356210
Also-by: Daniel Megert <daniel_megert@ch.ibm.com>
Change-Id: Id1c11283408ee37429dda794eea453e1d18344c3