[historyView] Remove status messages
The history view current displays a status message and hides
the commit graph table when loading or showing no input.  This
produces a flickering when switching between repositories or
when in link with selection mode.

Now the history view keeps the current table input while loading
the new input in the background and shows an empty table when
no commits are found to display.

The status message during loading did not seem necessary since
the job is present in the Progress view with information about
what is being loaded.  Also an empty table seems sufficient to
let the user know there are no commits to display.

Also no longer join the job when canceling the refresh job
since it can hang the UI thread.

Bug: 337399
Bug: 358650
Change-Id: I9a464431b99188f420b2ae1a7af1d2f3c8181916
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>