Allow push of single Refs/Commits from Repo & History Views
This adds a new SimplePushWizard which just allows selecting a target
remote and ref name. It then pushes out the selected Ref/Commit to the
selected target ref.

While implementing this I split out the Ref content assist functionality
from the existing RefSpecDialog to be reusable by this code.

Bug: 341076
Bug: 356314
Bug: 337747
CQ: 6347
Change-Id: If3be09886c23b031489116d448cae2dce29e281a
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>