[historyView] Use file viewer control when updating sash form
The file viewer is no longer wrapped in an extra layout and
so the sash form used at the bottom of the history view should
no longer use the parent control when setting the maximized
control.

This corrects an issue where the files in the selected commit
did not display in the history view when showing only the files
and not the files and comments both.

Change-Id: I7bd71231ebd249be8ab752bc2733a23e46b257f0
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>