[historyView] Allow to rename branches
It is often helpful to be able to rename a branch (for example, you
happened to commit into the "master" branch and want to rename it to
reflect your feature).

Without being able to do this on the History View, this is rather
cumbersome.

[MS]
Pre-populate the new name with the old name and select it fully. If
the user wants to just slightly alter the old name instead of replacing
it with a completely different name this requires less keystrokes for
the same result.

Change-Id: Ib059c8b00d9d062032dfa87400147ea80d364411
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>