Fix releasing RevWalks in GitHistoryPage
When a new History rendering was triggered a RevWalk is created. This
RevWalk was released when the next history rendering was triggered. 
However, at this point in time the RevWalk may still have been used by
the previous GenerateHistoryJob in another thread. This leads to NPEs.

Now we release the RevWalk when the GenerateHistoryJob is done.

Bug: 374675
Change-Id: I3685bd6994952b1378bbd14f39245d2edf64ee15
Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>
Signed-off-by: Stefan Lay <stefan.lay@sap.com>