Revert "Save Actions for removing braces of one-line statements"
This reverts commit f95c105980fc8dc1c356665bb796f256d62bf1ca.

The problem with the save action is that it also results in removed
braces for the if in the following example:

  if (foo)
    try {
      bar();
    } finally {
      baz();
    }

Which, altough it's a single statement, isn't recommended because it's
longer than one line and confusing.

See I1279030286ef177f3360ff70660a9365e4f3119a for the exact reasons.

Change-Id: I4564e3d312fe3663d22ac784de83408509275219
Signed-off-by: Robin Stocker <robin@nibor.org>