[historyView] Perform clear on background thread when disposed
An NPE can occur if SWTCommitList.clear() is called
from dispose() while fillTo() is running from
GenerateHistoryJob.

dispose() is always called on the UI-thread so schedule
a background thread that synchronizes on the instance
and calls clear since all other state accessing calls to
the SWTCommitList instance are done inside synchronized
blocks.

Change-Id: I855d5bfeb5131c634b4790e6b7231424f2be4566