[historyView] Handle branch deletion more user-friendly
Let's assume we have a commit which the currently checked out
branch and a remote tracking branch points to. With the current
implementation, right-clicking on "Delete Branch" will delete the remote
tracking branch quietly.

With this change, the branch selection dialog will come up offering to
delete the remote tracking branch.

Change-Id: Ia4096aef8f459cd3476f87fe9622b6dbefe29e03
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>