Warn user if Git's system wide settings cannot be found
A Git user may have important settings in the system wide configuration
file. That file is located relative to Git's installation directory.
EGit tries to find that file automatically by looking for the git
executable. If Git cannot be found then EGit also cannot find the system
wide settings.

In order to stay compatible with Git, in particular on Windows, where
Git may not appear in PATH, it is important that the user gets notified
that there may be a problem. The most important setting is typically
core.autocrlf.

Change-Id: Id1b6261edb778e82d76c0dfcfbf26228aa515ee9
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>